Create extract_all_page_view_metrics.sql
BQ is compliqué if you need specific metrics for a single event, but should this really be the case? this is the first of a number of queries to be pushed which creates a dedicated table for each event_name to do away with the nested fields to make life easier. this should help with reports in looker studio and think it should help  a lot for working with GA4 data in spreadsheets